+++ src/compiled/glsl/transpiler.rs
@@ -1069,16 +1069,19 @@ const HELPER_SDF_UNEVEN_CAPSULE: &str = r"float sdf_uneven_capsule(vec3 p, float
 ";
 
 const HELPER_SDF_EGG: &str = r"float sdf_egg(vec3 p, float ra, float rb) {
+    const float k = 1.7320508;
     float px = length(p.xz);
     float py = p.y;
     float r = ra - rb;
+    float d;
     if (py < 0.0) {
-        return length(vec2(px, py)) - r;
-    } else if (px * ra < py * rb) {
-        return length(vec2(px, py - ra));
+        d = length(vec2(px, py)) - r;
+    } else if (k * (px + r) < py) {
+        d = length(vec2(px, py - k * r)) - r;
     } else {
-        return length(vec2(px + rb, py)) - ra;
+        d = length(vec2(px + r, py)) - 2.0 * r;
     }
+    return d - rb;
 }
 ";